Fix operand types for various scalar crypto insns
A handful of the scalar crypto instructions are supposed to take a
constant integer argument 0..3 inclusive.  A suitable constraint was
created and used for this purpose (D03), but the operand's predicate is
"register_operand".  That's just wrong.

This patch adds a new predicate "const_0_3_operand" and fixes the
relevant insns to use it.  One could argue the constraint is redundant
now (and you'd be correct).  I wouldn't lose sleep if someone wanted
that removed, in which case I'll spin up a V2.

The testsuite was broken in a way that made it consistent with the
compiler, so the tests passed, when they really should have been issuing
errors all along.

This patch adjusts the existing tests so that they all expect a
diagnostic on the invalid operand usage (including out of range
constants).  It adds new tests with proper constants, testing the
extremes of valid values.

OK for the trunk, or should we remove the D03 constraint?

Jeff
PR target/110201
gcc/
	* config/riscv/predicates.md (const_0_3_operand): New predicate.
	* config/riscv/crypto.md (riscv_aes32dsi): Use new predicate.
	(riscv_aes32dsmi, riscv_aes32esi, riscvaes32esmi): Likewise.
	(riscv_sm4ed_<mode>, riscv_sm4ks_<mode): Likewise.

gcc/testsuite
	* testsuite/gcc.target/riscv/zknd32.c: Verify diagnostics are issued
	for invalid builtin arguments.
	* testsuite/gcc.target/riscv/zkne32.c: Likewise.
	* testsuite/gcc.target/riscv/zksed32.c: Likewise.
	* testsuite/gcc.target/riscv/zksed64.c: Likewise.
	* testsuite/gcc.target/riscv/zknd32-2.c: New test.
	* testsuite/gcc.target/riscv/zkne32-2.c: Likewise.
	* testsuite/gcc.target/riscv/zksed32-2.c: Likewise.
	* testsuite/gcc.target/riscv/zksed64-2.c: Likewise.