ada: Spurious error on package instantiation
The compiler reports spurious errors processing the instantation
of a generic package when the instantation is performed in the
the body of a package that has a private type T, a dispatching
primitive of T has the same name as a component of T, and
an extension of T is used as the actual parameter for a
formal derived type of T in the instantiation.

gcc/ada/

	* sem_ch4.adb
	(Try_Selected_Component_In_Instance): New subprogram; factorizes
	existing code.
	(Find_Component_In_Instance) Moved inside the new subprogram.
	(Analyze_Selected_Component): Invoke the new subprogram before
	trying the Object.Operation notation.

Tested on x86_64-pc-linux-gnu, committed on master.