ada: Fix type derivation of subtype of derived type
Deriving from a subtype of a derived type of a private type, whose full
view is itself a derived type of a discriminated record with a known
discriminatant was failing with the error message:

  invalid constraint: type has no discriminant

The compiler needs to use the full view to be able to constrain the
type.

Also fix minor typo in comments.

gcc/ada/

	* sem_ch3.adb (Build_Derived_Record_Type): Use full view as
	Parent_Base if needed.

Tested on x86_64-pc-linux-gnu, committed on master.