rs6000: Update the vsx-vector-6.* tests.
GCC maintainers:

Ver 2.  Switched to using code macros to generate the call to the
builtin and test the results.  Added in instruction counts for the key
instruction for the builtin.  Moved the tests into an additional
function call to ensure the compile doesn't replace the builtin call
code with the statically computed results.  The compiler was doing this
for a few of the simpler tests.

The following patch takes the tests in vsx-vector-6-p7.h,  vsx-vector-
6-p8.h, vsx-vector-6-p9.h and reorganizes them into a series of smaller
test files by functionality rather than processor version.

Tested the patch on Power 8 LE/BE, Power 9 LE/BE and Power 10 LE with
no regresions.

Please let me know if this patch is acceptable for mainline.  Thanks.

                       Carl

----------------------
rs6000: Update the vsx-vector-6.* tests.

The vsx-vector-6.h file is included into the processor specific test files
vsx-vector-6.p7.c, vsx-vector-6.p8.c, and vsx-vector-6.p9.c.  The .h file
contains a large number of vsx vector builtin tests.  The processor
specific files contain the number of instructions that the tests are
expected to generate for that processor.  The tests are compile only.

The tests are broken up into a seriers of files for related tests.  The
new tests are runnable tests to verify the builtin argument types and the
functional correctness of each test rather then verifying the type and
number of instructions generated.

gcc/testsuite/
	* gcc.target/powerpc/vsx-vector-6-1op.c: New test file.
	* gcc.target/powerpc/vsx-vector-6-2lop.c: New test file.
	* gcc.target/powerpc/vsx-vector-6-2op.c: New test file.
	* gcc.target/powerpc/vsx-vector-6-3op.c: New test file.
	* gcc.target/powerpc/vsx-vector-6-cmp-all.c: New test file.
	* gcc.target/powerpc/vsx-vector-6-cmp.c: New test file.
	* gcc.target/powerpc/vsx-vector-6.h: Remove test file.
	* gcc.target/powerpc/vsx-vector-6-p7.h: Remove test file.
	* gcc.target/powerpc/vsx-vector-6-p8.h: Remove test file.
	* gcc.target/powerpc/vsx-vector-6-p9.h: Remove test file.