Avoid shorten_binary_op on VECTOR_TYPE
When we disallow TYPE_PRECISION on VECTOR_TYPEs it shows that
shorten_binary_op performs some checks on that that are likely
harmless in the end.  The following bails out early for
VECTOR_TYPE operations to avoid those questionable checks.

Bootstrap and regtest ongoing on x86_64-unknown-linux-gnu.

OK?

gcc/c-family/
	* c-common.cc (shorten_binary_op): Exit early for VECTOR_TYPE
	operations.