d: Suboptimal codegen for __builtin_expect(cond, false)
Hi,

Since PR96435, both boolean objects and expressions have been evaluated
in the following way by the D front-end.

    (*(ubyte*)&obj_or_expr) & 1

It has been noted that sometimes this can cause the back-end to optimize
in non-obvious ways - in particular with __builtin_expect.

This @safe feature is now restricted to just when reading the value of a
bool field that comes from a union.

Bootstrapped and regression tested on x86_64-linux-gnu/-m32, committed
to mainline, and backported to releases/gcc-13 through to gcc-10.

Regards,
Iain.