ada: Fix too late finalization and secondary stack release in iterator loops
Sem_Ch5 contains an entire machinery to deal with finalization actions and
secondary stack releases around iterator loops, so this removes a recent
fix that was made in a narrower case and instead refines the condition under
which this machinery is triggered.

As a side effect, given that finalization and secondary stack management are
still entangled in this machinery, this also fixes the counterpart of a leak
for the former, which is a finalization occurring too late.

gcc/ada/

	* exp_ch4.adb (Expand_N_Quantified_Expression): Revert the latest
	change as it is subsumed by the machinery in Sem_Ch5.
	* sem_ch5.adb (Prepare_Iterator_Loop): Also wrap the loop
	statement in a block in the name contains a function call that
	returns on the secondary stack.

Tested on x86_64-pc-linux-gnu, committed on master.